bulk checkpoint of r&d.
This is more than a few days old already, so unfortunately I've lost a
lot of the mental context I'd like to pour into this commit message.

The big theme here is: I started porting filesystem composition
features based on doing tons and tons of my own mounting logic...

and...

... found out maybe I can escape that.  Runc can do a lot of things;
it can do a fair number of them rootlessly (!) (if... with some
amount of prancing around, admittedly); and so if we can do that,
we should.

There's a metric *ton* of edge cases around that, though.
The linux design of mounts needing directories under their target
actually singlehandedly makes a shocking amount of hellscape.
(You wanted to compose readonly filesystems?  Yeah have fun with that.)

So, a lot of this is still experimental.
Most of the code here is just composing and templating config,
and doesn't actually invoke it yet.
But it's heading in what seems to be a feasible direction.