Save some filesystem composition code.
This is from a few moons ago.  Its origins are even older: it's
largely extraction and simplification of code from either repeatr
or rio (I don't remember which).

I'm going to merge-ignore this.  I think we do not actually need it!
Using the mount composition system from within the container systems
gets us results without higher levels of privs on the host;
that's winning.