Thoughts on error handling.
It's been a buggaboo in every sufficiently large golang program
I've ever written.  Here's my latest take on it.  It's not unlike
the https://github.com/warpfork/go-errcat/ concept, but gives up
on bothering to have an interface -- it's just a struct, so that
it's explicitly and obviously serializable (and deserializable).

We'll see how it goes.

I still feel like one could borderline use a small codegen here.
The categories ought to be in a constants table.
The messages would be nicer to write if we could use kwargs in them.
The function args are roughly duplicates of the details mapping.
The boilerplate quotient here is... generally high.
Maybe there's something interesting that could be done.

But I'm going to resist depth-firsting on this.
Golang error handling is just a "known problem" to me now.