Begin outline of workspace package, and the filesystem detection.
This is going to look and feel a lot like where we left off with the
last round of timeless stack development, but with one fairly major
distinction: working *without* a workspace is no longer a concept.

I've also added a bit more distinctiveness to workspaces in the user's
homedir.  Actual functional distinctions on this aren't seen yet,
but I think we're going to.