Tests for workspace discovery.
Using the new fstest package from stdlib!  Nice!

... sort of.  It doesn't seem to work if you use paths that have a
leading slash.  Which on the one hand, I understand the allure of.
On the other hand: it creates some awkward edge cases for us, because
we very much have code that's gonna start with rooted paths by default,
and so this pushes some rather odd looking normalization onto us.
And I'm pretty sure error messages somewhere are going to get stranger
as a result of this, which doesn't thrill me.

As well; we'll debug those things as they continue to come, I guess.

Added quicktest to the deps list.

Also had to change some critical things about the errors standard the
very first instant it met use, heh.  Typed nils are a big problem,
around errors in particular for some reason.