Start drafting API/serial types for catalogs.
Describing them with IPLD schemas, and the go-ipld-prime bindnode tools
for binding to handwritten golang structs.  (Bumped go-ipld-prime
library version in order to get the bindnode stuff, which is recent.)

You can see some cool outcomes from leaning on that already:
for example, we got to use a stringjoin representation on a struct,
and that didn't require additional artisinal parsing code unique to
this project, which is awesome.

Hopefully the amount of boilerplate for this setup continues to trend
down in the future; right now all the IPLD schema stuff is specified
using a golang-DSL-ish constructor boilerplate monstrosity, and what
we'd like to do is just use the actual nice human-readable and
non-golang-specific schema DSL syntax.  That's not quite ready yet in
the upstream, though.

Unions are also still a bit bleigh to work with using bindnode.
That's... gonna be kinda a blocker.  Will be making upstream issue.