The full complexity of formula structures and parsing.
This time, we're leaning in even harder on stringprefix unions to
declare what's going on... and (though this isn't visible here yet)
one of the idea of this is to keep the same prefixes consistent
across layers (which is why it says "ware" at this layer, despite
that arguably seeming like a "well duh" sort of thing).

I'm also rolling with the idea that we'll do environment variables
as inputs this time.  (This might even be generalized to other
concepts of "variables" for non-container computations... in the
very distant future, though; that's speculative and very non-MVP).

Actions still need more filling it, but there won't be any big
surprises there, I think.

FormulaContext is also included here.  It's much briefer than it was
in previous generations.  We only do warehouses for the input fetching
phase of data flow, now.  There's no need for warehouse config for
the output side because we just always, always use a local warehouse
for that.  (Failing to emit something to the network at the end of
a long build is an awful experience; let's not have it anymore.)
The local warehouse will be configurable in the nearest workspace
flagged as a root; that should do nicely.

More notes and docs are in Notion.  (It might not be as well-versioned,
but I'm finding it much easier to write and edit there, and it's
also a much shorter feedback loop to collaborate with others than
text files in git, so it's getting more of my attention lately.)

There's a bump to go-ipld-prime here because some fixes to various
deserialization features we're using here (mostly around unions)
were needed upstream.

Test fixture included.  It's nonsense data, but close enough to the
real thing one can get a feel for how it will be used, I think.