This is what computing CIDs looks like.  Parse fixture system can now check these too.
This may be non-final.  We might change the way we string these a bit.
I'd also like to use a sha384, which is currently not wired up for
sufficiently trivial use in the upstream gadgetry (and it really
should be).

May also need some more helper functions about this in the upstream.
The verbosity is pretty nuts here.
(But we can write a very small number of helper functions ourselves
in this project, too, and probably get over it very quickly.
We'll only need to do this for like... less than a half dozen different
kinds of documents, I think; and that helper function will actually be
naturally generic too because of the whole "node" thing.)