larkery: successful first creation of a struct.
Nice.

The way these symbols are getting foisted into starlark is placeholder,
of course.  Time soon to get a hold of how to feed in dotted notation.

(I'm not confident in what the right choice is for binding receivers,
or if binding receivers is even... yeah.  More notes on this soon.)