larkery: decision on a strategy for prototypes.
We *can* do Callable ourselves, and it works *awesome*.  Sweet.

The existing tests were easy to keep working.
Bonus?  They actually got a *lot* shorter to wire up, because
the Callable interface turns out to be a snugger fit for what we
actually need than the kind of function that NewBuiltin asks for.

Also bonus?  The debug printout names got more informative, as hoped.