larkery: maps construct in standard way now.  Various other standardizations.
This is feeling increasingly reasonable.

Mostly, the naming is what needs to get hammered out now.
And something feels a little awkward around untyped stuff
(in particular, whether it's even appropriate to have named functions
for that?  versus just having it "DTRT" based on the starlark values).
I think the next step will be cleaning up the examples and making more
of them so we can start seeing real syntactical outcomes.