larkery: map constructors can now accept and transform a dict; also, our strings are hashable (meaning they can be used as keys).
This finally lets me uncomment a line in an example that does...
very interesting things.

It'll get more interesting when there's more types in flight.

(Right now, it's a bit of a dummy operation, still; it's just a
constructor call that has to deal with typed keys instead of bare
regular starlark strings.  When we get things that aren't even
exactly strings though?  That's gonna be neat.)