improve tests and examples to use testexec's tmpdirs and file fixture features.
The upside is that the examples are now showing everything, again,
in one view -- so as a reader, you don't have to jump your eyes to
multiple other files in other documents.

Also, using hunks of the "fs/*" naming pattern means testexec is
making tmpdirs for us, and that significantly improves our isolation.

The improved isolation also points out at least one thing currently
broken, though: these examples were relying on the warpforge repo's
workspace files to be readable, and have this particular catalog data.
We shouldn't be doing that; we should be establishing those fixture
files in the same place, too: so that it's isolated; and so that
a casual reader can *see* it, so these files have explanatory value.

There's also probably now some reorganizing work we could do to make
this example file more readable, now that everything is visible.