catalogsite: checkpoint of a (skeletal) catalog html generator.
Containing copious hacks in the "test".  Not production-ready.
But still I still may want to merge this just to move forward.

The exact filesystem layout may be subject to change.
This is a rebase of an old prototype (from before warpforge v0.2);
the exact filesystem I'd like to use nowaday should more closely
mirror the paths of the catalog filesystem projection in JSON,
just with different extensions (which is not what we have here).

Everything else _also_ subject to change :)  There's no styling, etc.

Future work includes turning this into a real CLI command;
giving it parameters for which catalog(s?) to generate content from;
teaching it to render and link up certain pieces of metadata
(especially, right now, Replays);
and probably a lot more.