Fix some regressions from the conversions to using fs.FS.
Turns out fs.FS is very spikey about receiving a param that begins with
a slash.  Kind of unfortunate, honestly.  Would be a lot easier to use
if it just decided to normalize that out.  Ah well.

(Also turns out one of these was harder to notice than it should've
been because we were shrugging at an over-wide range of errors, rather
than only shrugging at a specifically acceptable not-exists error...
so that's a bug fixed, coincidentally.)

I hope this fixes all regressions.  It's not the cleanest changes now,
either... but this should continue to improve if we get writablity
features attached to the fs package in the future (which, again,
hopefully is very (very) soon now).