ComputeStats on a plot needs to be recursive!
The actual code that evaluates and execs things is, so the estimator
most definitely must match that as well.

Previously, this would underestimate how many things a plot was
referencing, and could miss it if mounts or ingests were used in a
subplot or in a protoformula instead of being imported and piped in.

If we had been using this for a security boundary, I'd say this is
rather bad.  However, we've only been treating this data as advisory
for now anyway, so... perhaps no klaxons.  But yikes.