'workspace inspect' command now generates many checkmarks.
Okay, not checkmarks yet.  I'm going to prettify this later.

But it does a bunch of interesting accounting.

And also, yes, having the ability to say "--gohard=false" is already
relevant in its performance impacts.  In the 29 modules currently
in the warpsys workspace... it takes 0m0.008s real wallclock time
to just enumerate them (makes sense; it's mostly walk, and just open
and parse on the module file itself to get the name)...

... and it takes a whopping 0m0.390s real wallclock seconds to collect
all these connectivity and health checks.  Like, it's *noticable*.

Mind, I suspect there's very low-hanging fruit for improving that.
I notice the system time is 0m0.273s of that (and 0m0.133s user time
from the same sample), so... I suspect even just throwing some caching
around the reads of the catalog would probably make a big difference.

Tracing might tell us more (but that work is going on parallel to this,
so, more on that later perhaps).