'workspace inspect' now has valence opinions on these checkboxes.
Characters placeholder.  I'm being entirely silly here.

This should get replaced by a JSON API or something, and a filter
function that emits ANSI colors and such.

But!  The demo is neat.  If you pipe it to "column -t", it DTRT,
and when using it on the warpsys repo?  Hej, suddenly all the example
modules stand out... because they use ingests, and get flagged with
cautionary markers!  Nice.