fix up cli tests
With the new additions to testmark, we have a bunch of tests to fix that
weren't actually executing before. This gets us back into a relatively
good state where all the tests pass.

This change also makes significant changes to how the offline "/net/"
flag works. I.E. Any non-executable children of the parent DirEnt are
copied to the "/net/" DirEnt before execution. This allows a parent with
both online and offline tests to execute and properly pass along
required filesystems, although the parent script/sequence may execute
twice. However, this is _not required_ and a parent of a "/net/"
directory does not need to be executable or have any other children if
the "/net/" DirEnt has executable children. The children copied to the
"/net/" DirEnt must be unique. Any duplicates will be an error.