Massively refactored CLI wiring.
The main justifications of this are to make it easier to test,
and to prepare for more features I want to work on that require
more access to the CLI objects.

The justifications for this package structure can be found in
the introduced HACKME file.

The use of package-scope vars for the CLI object is considerably unfortunate,
but (again, as detailed in the HACKME), our CLI library already breaks the dam.
(This will become even more largely applicable when I get to work on
the planned features around help content generation, which will require
yet more global var touches to things inside the CLI library.)
So: in for a penny, in for a pound, and I hope I don't regret this too much.

Subcommands now register themselves in package init mechanisms.
As above: in for a penny, in for a pound, and I hope I don't regret this too much.

There's no more use of "internal" packages.
Those tend to footgun me painfully more often than not,
and so let's not have that.

Removed a fairly bonkers use of an "after" func in the CLI.
Its only user was part of the check subcommand.
Fewer lines of code and fewer weird long jumps across callbacks
occur by simply putting the code where it's used.

Many other miscellaneous fixes.
For example, fixing many references to `os.Args[0]` to use the app name.
Too many other small ones to recount.

A few things itch to me, but haven't yet received any changes:
there's still a "util" package in 'app/base/util',
and I'd like to tidy it away.
Also, the "middleware" functions in it for tracing are highly
redundant and I wonder if we can't make that a bit less frequently
copy-pasted.
This diff is already big enough, though.

The 'pkg/config' package also continues to exist, and I think it
should probably be unified into 'app/base' shortly.
The CLI library has features for env vars, as well as documenting them,
and we might as well use that.
(This config package is also pretty brazenly using package-scope vars
already too, so combining it with this other package that also does so
should produce no additional hurt.)

There are some uses of "../" in test setup that are also likely to
become broken soon -- as soon as we use 'app/testutil' from varying
depths of packages.  In this commit, that hasn't happened yet,
but expect some minor Fun dealing with that soon.