Move watch and spark packages to pkg folder.
The "subcmd" folder seems confusing.
We now have packages per CLI command group,
and we have packages that are meant for reuse,
and I don't know exactly where these pieces of code are on
that spectrum, but I don't think a third whole category
will help anyone else understand it either, and is in fact
as likely to add confusion as not.

The "pkg/healthcheck" package already found itself in a
similarly ambiguous situation, and was already placed similarly,
so this continues that convention.

I *would* accept an argument that packages that contain
daemon-like code and goroutine usage belong in a different
grouping than code that's just directly callable...
and if we made that organizational ruling, then *some* of
the code in these packages might move to match.
But that's not today.

So.  For now: One less subtree.  Less to think about.