Update all dependencies.
The CLI help text output changed, very slightly, for the better.
A fixture is updated accordingly.

The name of a field in the IPLD library changed.  A small rename.
The new name makes much more sense, so fair enough.

The dumbest change is a package import in the opentelemetry stuff.
I had to guess the value "v1.17.0" -- and I do mean **guess**;
at the moment, the latest thing in that folder is "v1.18.0",
and I have no idea what any of these are or how I'd choose.
If you don't guess right, you just get the error message of
"cannot merge resource due to conflicting Schema URL", which is
fairly remarkable unhelpful.  I have no idea what this dance
is supposed to accomplish, because it appears to me that the
number must match something within other parts of the upstream.
Ah well.  I guessed until I found the magic number.  Moving on.

That was it; less bumpy than one might fear for a big dep bump.