Checkpoint of taking control of CLI help generation.
Minimal changes vs the upstream so far, other than
cosmetically in the code (I added a helper for heredoc indents),
and a few whitespace tweaks in the output just to prove I can.

This commit just gets it working.  Though it is a bit large,
I think this was about the minimal level of invasion possible.

After this, I think we can start getting more exciting.