Replace one of the main templates with markdown, and reformat it.
Golang's text template system still doesn't make whitespace easy to handle clearly,
but in this case I think legibilty already went up in the templates,
and I got approximately the same result.  (There's a few more breaks than desired
right under the headings; I may or may not chase those.)

I'm moving on to experimenting with attaching markdown processing to the end of this,
which may make a lot of the whitespace conversation change or become moot anyway,
so I won't spend much more time on it here.