Research checkpoint: attempting to use lipgloss and goldmark.
... and I think, after playing with it for a good number of hours,
this ain't it either.

lipgloss is coming in at a really weird middle level of abstraction.
The features like margins and wrapping aren't really doing the right
things without an incredible amount of cajoling (and even then).
It seems to have kind of _half_ a component model, and doesn't seem
to compose very gracefully -- it looks like I'll just *keep* passing
width info down on the side, while I use it, which is confusing.
I'm not sure if I'm holding it wrong or if it just doesn't have the
same view as me on what's important.

And it seems to have the helpful opinion that it's gonna emit
CSI reset codes constantly... which absolutely does not compose
with how goldmark AST visitors (or any sane AST visit system) is
going to want to get things done.

Less importantly, but perturbing to me, is that lipgloss is also doing
just a staggering amount of map lookups and memory allocations for
every little thing.

All in all, I think I'm going to have to dive another level deeper,
because this just isn't doing the thing.

(I'm halfway tempted to pull up and go back to using glamour, because
that got an interesting MVP radically faster, but... it just plateaus
so dang hard.  I can't do variable indent; the tables were meh; the
options for doing anything transformative with links was ziltch...
No, I don't think I really wanna rest there either.)

I'm thinking just a few key functions for wrapping text with awareness
of ANSI codes and their (lack of) effect on render size will maybe
be entirely sufficient and do a simpler better job.
A much simpler library (or even just set of short functions; ANSI codes
are distinctly Not That Complicated) for the highlighting will also
totally suffice.  So: that'll be the approach to try up next.