Drop glamour and lipgloss experiments; ansi ourselves.
I think we've finally got up to the glamour demo's equivalent,
but also with control over the other things we wanted.

(I'm beginning to regret not simply doing this a LONG time ago, with
drastically fewer indirections, and perhaps not even the markdown bit.
This journey has been FAR longer and sillier than I expected.)

Wordwrap finally works right.  Paragraphs are indented according to
the level of the heading that precedes them.  Finally.

The reflow library works fairly well for the wordwrap.
No complaints other than it's a little complicated in its guts.
And that I'm a bit sad I can't indent in the same pass as wrapping.
But those are small potatoes issues and probably not worth rewriting
that dang thing for, too.

The styling of inline elements within other nodes like paragraphs
is not yet honored; that's still a todo.

I still haven't introduced any links into the markdown, nor
processing for them either here or in the website, but those are
now an obviously reachable distance from this checkpoint.

In the brief custom ANSI code wrangling I banged out, here:
I don't love it, but this honestly seems simpler than not owning it.
I've only bothered to support a few well-known colors in the table
so far, but it's sufficient to proceed with.  Some structs might
be in order for the future if we support rgba stuff; that can come
when it's needed (or, maybe we can look again and have luck finding
a library that does the essentials and not too wildly much more).

I'm dropping glamour and lipgloss from the experients entirely now,
and go.mod shrinks considerably, accordingly.