Alright, let's go.  New markdown rendering attached.
In the real CLI, it uses the "ansidown" mode.
(I've hardcoded it for now; but surely, it should do some detection
of feature and situation, and be disableable.  Future work.)

In the tests that regenerate the website content: it uses plain
markdown mode.  This gets the content through in a nice form that
let the website pipeline handle redecoration.  And passing through
the whole parser gets us some whitespace regularlization.