Refactor the run command's search for files.
It now searches for everything up-front.

My main goal with this was to get towards having a list of what all
we're going to do, so that we can make progress reports about how far
along we are.  The second half of that remains a long road ahead,
but the first part is now significantly closer.

It's now detected faster if you typo a filename, rather than
potentially being noticed only after lots of other work has been done,
which is a mild improvement.

It also supports more than one "..." pattern now, although that was
sort of a target-of-opportunity thing.

Overall, this isn't a huge change itself, but cleans things up and
separates some concerns nicely, and prepares they way for more.

Bumps the go-fsx version, to get some more convenience functions.
(I'm also starting to slowly increase our use of FS objects.)